Move oscillator constructs into their own place
