+++ internal/oscillator/oscillator.go
@@ -1,4 +1,4 @@
-package util
+package oscillator
 
 import (
 	"math/rand"
